+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - GAP: simulate HCI_EVENT_REMOTE_NAME_REQUEST_COMPLETE if HCI Remote Name Request fails 
 - AVRCP Controller: Added send generic PASS THROUGH command
+- HID Host: storage for HID Descriptors is now optional
 ### Fixed
 - GAP: store link key for standard/non-SSP pairing
 ### Changed